refactor: split expression type inference
