+++ src/types/checker/inference/expr/mod.rs
@@ -12,12 +12,12 @@ use crate::errors::CompileError;
 use crate::names::php_symbol_key;
 use crate::parser::ast::{Expr, ExprKind};
 use crate::span::Span;
-use crate::types::{
-    merge_array_key_types, normalized_array_key_type, packed_type_size, PhpType, TypeEnv,
-};
+use crate::types::{PhpType, TypeEnv};
 mod assignments;
+mod basic;
 mod class_refs;
 mod effects;
+mod calls_objects;
 mod static_closure;
 use super::super::Checker;
 use super::syntactic::null_coalesce_merge_type;
@@ -34,674 +34,69 @@ impl Checker {
     /// validation, and optimizer-visible type metadata.
     pub fn infer_type(&mut self, expr: &Expr, env: &TypeEnv) -> Result<PhpType, CompileError> {
         match &expr.kind {
-            // `IncludeValue` is a transient parser node fully expanded by the resolver;
-            // it can never reach this pass.
-            ExprKind::IncludeValue { .. } => unreachable!(
-                "ExprKind::IncludeValue must be expanded by the resolver"
-            ),
-            ExprKind::BoolLiteral(false) => Ok(PhpType::False),
-            ExprKind::BoolLiteral(true) => Ok(PhpType::Bool),
-            ExprKind::Null => Ok(PhpType::Void),
-            ExprKind::StringLiteral(_) => Ok(PhpType::Str),
-            ExprKind::IntLiteral(_) => Ok(PhpType::Int),
-            ExprKind::FloatLiteral(_) => Ok(PhpType::Float),
-            ExprKind::Variable(name) => self.variable_type_or_eval_dynamic(name, expr.span, env),
-            ExprKind::Negate(inner) => {
-                let ty = self.infer_type(inner, env)?;
-                match ty {
-                    PhpType::Int => {
-                        if matches!(&inner.kind, ExprKind::IntLiteral(_)) {
-                            Ok(PhpType::Int)
-                        } else {
-                            Ok(PhpType::Mixed)
-                        }
-                    }
-                    PhpType::Float => Ok(PhpType::Float),
-                    PhpType::Mixed | PhpType::Bool | PhpType::False | PhpType::Void => {
-                        Ok(PhpType::Mixed)
-                    }
-                    _ => Err(CompileError::new(
-                        expr.span,
-                        "Cannot negate a non-numeric value",
-                    )),
-                }
-            }
-            ExprKind::Not(inner) => {
-                self.infer_type(inner, env)?;
-                Ok(PhpType::Bool)
-            }
-            ExprKind::ErrorSuppress(inner) => self.infer_type(inner, env),
-            ExprKind::Print(inner) => {
-                self.infer_type(inner, env)?;
-                Ok(PhpType::Int)
-            }
-            ExprKind::PreIncrement(name) | ExprKind::PreDecrement(name) => match env.get(name) {
-                Some(PhpType::Int) => Ok(PhpType::Mixed),
-                Some(PhpType::Mixed) => Ok(PhpType::Mixed),
-                Some(PhpType::Bool) | Some(PhpType::False) | Some(PhpType::Void) => {
-                    Ok(PhpType::Int)
-                }
-                Some(other) => Err(CompileError::new(
-                    expr.span,
-                    &format!("Cannot increment/decrement ${} of type {:?}", name, other),
-                )),
-                None => Err(CompileError::new(
-                    expr.span,
-                    &format!("Undefined variable: ${}", name),
-                )),
-            },
-            ExprKind::PostIncrement(name) | ExprKind::PostDecrement(name) => match env.get(name) {
-                Some(PhpType::Int)
-                | Some(PhpType::Bool)
-                | Some(PhpType::False)
-                | Some(PhpType::Void) => Ok(PhpType::Int),
-                Some(PhpType::Mixed) => Ok(PhpType::Mixed),
-                Some(other) => Err(CompileError::new(
-                    expr.span,
-                    &format!("Cannot increment/decrement ${} of type {:?}", name, other),
-                )),
-                None if self.eval_barrier_active => Ok(PhpType::Int),
-                None => Err(CompileError::new(
-                    expr.span,
-                    &format!("Undefined variable: ${}", name),
-                )),
-            },
-            ExprKind::ArrayLiteralAssoc(pairs) => {
-                if pairs.is_empty() {
-                    return Err(CompileError::new(
-                        expr.span,
-                        "Cannot infer type of empty associative array literal",
-                    ));
-                }
-                let mut key_ty = normalized_array_key_type(
-                    &pairs[0].0,
-                    self.infer_type(&pairs[0].0, env)?,
-                );
-                let mut val_ty = self.infer_type(&pairs[0].1, env)?;
-                for (k, v) in &pairs[1..] {
-                    let kt = normalized_array_key_type(k, self.infer_type(k, env)?);
-                    let vt = self.infer_type(v, env)?;
-                    if kt != key_ty {
-                        key_ty = merge_array_key_types(key_ty, kt);
-                    }
-                    if vt != val_ty {
-                        val_ty = PhpType::Mixed;
-                    }
-                }
-                Ok(PhpType::AssocArray {
-                    key: Box::new(key_ty),
-                    value: Box::new(val_ty),
-                })
-            }
-            ExprKind::Match {
-                subject,
-                arms,
-                default,
-            } => {
-                self.infer_type(subject, env)?;
-                let mut result_ty: Option<PhpType> = None;
-                for (conditions, result) in arms {
-                    for c in conditions {
-                        self.infer_type(c, env)?;
-                    }
-                    let ty = self.match_arm_result_type(result, env)?;
-                    result_ty = Some(match result_ty {
-                        Some(acc) => merge_match_arm_result_type(self, acc, ty),
-                        None => ty,
-                    });
-                }
-                if let Some(d) = default {
-                    let ty = self.match_arm_result_type(d, env)?;
-                    result_ty = Some(match result_ty {
-                        Some(acc) => merge_match_arm_result_type(self, acc, ty),
-                        None => ty,
-                    });
-                }
-                Ok(result_ty.unwrap_or(PhpType::Void))
-            }
-            ExprKind::ArrayLiteral(elems) => {
-                if elems.is_empty() {
-                    return Ok(PhpType::Array(Box::new(PhpType::Never)));
-                }
-                if elems.iter().any(|elem| {
-                    matches!(
-                        &elem.kind,
-                        ExprKind::Spread(inner)
-                            if matches!(
-                                self.infer_type(inner, env),
-                                Ok(PhpType::AssocArray { .. })
-                            )
-                    )
-                }) {
-                    let value_ty = self.assoc_spread_literal_value_type(elems, env);
-                    return Ok(PhpType::AssocArray {
-                        key: Box::new(PhpType::Mixed),
-                        value: Box::new(value_ty),
-                    });
-                }
-                let mut elem_ty = self.infer_type(&elems[0], env)?;
-                for elem in &elems[1..] {
-                    let ty = self.infer_type(elem, env)?;
-                    if ty != elem_ty {
-                        if let Some(merged_ty) = self.merge_array_element_type(&elem_ty, &ty) {
-                            elem_ty = merged_ty;
-                            continue;
-                        }
-                        elem_ty = PhpType::Mixed;
-                    }
-                }
-                Ok(PhpType::Array(Box::new(elem_ty)))
-            }
-            ExprKind::ArrayAccess { array, index } => {
-                let arr_ty = self.infer_type(array, env)?;
-                let idx_ty = self.infer_type(index, env)?;
-                let normalized_idx_ty = normalized_array_key_type(index, idx_ty.clone());
-                match &arr_ty {
-                    PhpType::Str => {
-                        if !is_valid_string_offset_index(index, &idx_ty) {
-                            return Err(CompileError::new(
-                                expr.span,
-                                "String index must be integer",
-                            ));
-                        }
-                        Ok(PhpType::Str)
-                    }
-                    PhpType::Array(elem_ty) => {
-                        if normalized_idx_ty != PhpType::Int {
-                            // PHP allows string keys on indexed arrays: the array
-                            // promotes to hash at runtime. Return the element type
-                            // widened to Mixed so ?? / isset / reads type-check.
-                            Ok(PhpType::Mixed)
-                        } else {
-                            Ok(*elem_ty.clone())
-                        }
-                    }
-                    PhpType::AssocArray { value, .. } => {
-                        // Assoc arrays accept string or int keys
-                        Ok(*value.clone())
-                    }
-                    PhpType::Object(class_name) => {
-                        if self.object_type_implements_interface(class_name, "ArrayAccess") {
-                            Ok(self.array_access_offset_get_type(class_name))
-                        } else {
-                            Err(CompileError::new(expr.span, "Cannot index non-array"))
-                        }
-                    }
-                    PhpType::Union(members) => {
-                        let mut result_members = Vec::new();
-                        let mut saw_indexable_member = false;
-                        let mut first_index_error = None;
-                        for member in members {
-                            match member {
-                                PhpType::Void => result_members.push(PhpType::Void),
-                                PhpType::Str => {
-                                    saw_indexable_member = true;
-                                    if !is_valid_string_offset_index(index, &idx_ty) {
-                                        first_index_error =
-                                            first_index_error.or(Some("String index must be integer"));
-                                        continue;
-                                    }
-                                    result_members.push(PhpType::Str);
-                                }
-                                PhpType::Array(elem_ty) => {
-                                    saw_indexable_member = true;
-                                    if normalized_idx_ty != PhpType::Int {
-                                        // String key on indexed array: PHP promotes
-                                        // to hash at runtime; element may be Mixed.
-                                        result_members.push(PhpType::Mixed);
-                                    } else {
-                                        result_members.push(*elem_ty.clone());
-                                    }
-                                }
-                                PhpType::AssocArray { value, .. } => {
-                                    saw_indexable_member = true;
-                                    result_members.push(*value.clone());
-                                }
-                                PhpType::Object(class_name) => {
-                                    if self.object_type_implements_interface(
-                                        class_name,
-                                        "ArrayAccess",
-                                    ) {
-                                        saw_indexable_member = true;
-                                        result_members
-                                            .push(self.array_access_offset_get_type(class_name));
-                                    }
-                                }
-                                PhpType::Buffer(elem_ty) => {
-                                    saw_indexable_member = true;
-                                    if !matches!(idx_ty, PhpType::Int | PhpType::Mixed) {
-                                        first_index_error =
-                                            first_index_error.or(Some("Buffer index must be integer"));
-                                        continue;
-                                    }
-                                    match elem_ty.as_ref() {
-                                        PhpType::Packed(name) => {
-                                            result_members.push(PhpType::Pointer(Some(name.clone())))
-                                        }
-                                        _ => result_members.push(*elem_ty.clone()),
-                                    }
-                                }
-                                _ => {}
-                            }
-                        }
-                        let has_concrete_result =
-                            result_members.iter().any(|member| *member != PhpType::Void);
-                        if !has_concrete_result && saw_indexable_member {
-                            Err(CompileError::new(
-                                expr.span,
-                                first_index_error.unwrap_or("Cannot index non-array"),
-                            ))
-                        } else if result_members.is_empty() {
-                            Err(CompileError::new(expr.span, "Cannot index non-array"))
-                        } else {
-                            Ok(self.normalize_union_type(result_members))
-                        }
-                    }
-                    PhpType::Buffer(elem_ty) => {
-                        if !matches!(idx_ty, PhpType::Int | PhpType::Mixed) {
-                            return Err(CompileError::new(
-                                expr.span,
-                                "Buffer index must be integer",
-                            ));
-                        }
-                        match elem_ty.as_ref() {
-                            PhpType::Packed(name) => Ok(PhpType::Pointer(Some(name.clone()))),
-                            _ => Ok(*elem_ty.clone()),
-                        }
-                    }
-                    // Mixed receivers fall through to runtime dispatch. The
-                    // boxed payload may carry an indexed array, an assoc
-                    // hash, or a stdClass; codegen unboxes and routes to
-                    // the right runtime helper. Missing keys decode to
-                    // `Mixed(null)` at runtime, mirroring PHP's silent
-                    // "undefined index" warning behavior for this very
-                    // common idiom (e.g. `json_decode($json, true)["k"]`).
-                    PhpType::Mixed => Ok(PhpType::Mixed),
-                    _ => Err(CompileError::new(expr.span, "Cannot index non-array")),
-                }
-            }
-            ExprKind::Ternary {
-                condition,
-                then_expr,
-                else_expr,
-            } => {
-                self.infer_type(condition, env)?;
-                // Flow-narrowing across the branches: `$x instanceof X ? ... : ...` (and the other
-                // recognized guards) narrow `$x` — or a simple `$x->prop` — in the then/else
-                // branches. A ternary is a single expression (no intervening writes), so the
-                // narrowing is safe to scope to each branch's inference.
-                let (then_ty, else_ty) = if let Some(guard) =
-                    self.guard_narrowing(condition, env)?
-                {
-                    let mut then_env = env.clone();
-                    then_env.insert(guard.var.clone(), guard.then_ty);
-                    let mut else_env = env.clone();
-                    else_env.insert(guard.var, guard.else_ty);
-                    (
-                        self.match_arm_result_type(then_expr, &then_env)?,
-                        self.match_arm_result_type(else_expr, &else_env)?,
-                    )
-                } else {
-                    (
-                        self.match_arm_result_type(then_expr, env)?,
-                        self.match_arm_result_type(else_expr, env)?,
-                    )
-                };
-                // Same Mixed/nullable merge as match arms: heterogeneous heap
-                // types must not collapse through the Str-absorbing syntactic join.
-                Ok(merge_match_arm_result_type(self, then_ty, else_ty))
-            }
-            ExprKind::ShortTernary { value, default } => {
-                let value_ty = self.match_arm_result_type(value, env)?;
-                let default_ty = self.match_arm_result_type(default, env)?;
-                Ok(merge_match_arm_result_type(self, value_ty, default_ty))
-            }
-            ExprKind::Throw(inner) => {
-                let thrown_ty = self.infer_type(inner, env)?;
-                match thrown_ty {
-                    PhpType::Object(type_name)
-                        if self.object_type_implements_throwable(&type_name) =>
-                    {
-                        Ok(PhpType::Void)
-                    }
-                    PhpType::Object(_) => Err(CompileError::new(
-                        expr.span,
-                        "Type error: throw requires an object implementing Throwable",
-                    )),
-                    _ => Err(CompileError::new(
-                        expr.span,
-                        "Type error: throw requires an object value",
-                    )),
-                }
-            }
-            ExprKind::Cast { target, expr } => {
-                self.infer_type(expr, env)?;
-                use crate::parser::ast::CastType;
-                Ok(match target {
-                    CastType::Int => PhpType::Int,
-                    CastType::Float => PhpType::Float,
-                    CastType::String => PhpType::Str,
-                    CastType::Bool => PhpType::Bool,
-                    CastType::Array => PhpType::Array(Box::new(PhpType::Int)),
-                })
-            }
-            ExprKind::FunctionCall { name, args } => {
-                let name = name.as_str().to_string();
-                let args = args.clone();
-                if self.extern_functions.contains_key(name.as_str()) {
-                    return self.check_extern_function_call(name.as_str(), &args, expr.span, env);
-                }
-                if let Some(ty) = self.check_builtin(name.as_str(), &args, expr.span, env)? {
-                    self.builtin_call_types.insert(expr.span, ty.clone());
-                    return Ok(ty);
-                }
-                self.check_function_call(name.as_str(), &args, expr.span, env)
-            }
-            ExprKind::BufferNew { element_type, len } => {
-                let len_ty = self.infer_type(len, env)?;
-                if len_ty != PhpType::Int {
-                    return Err(CompileError::new(
-                        expr.span,
-                        "buffer_new<T>() length must be integer",
-                    ));
-                }
-                let elem_ty = self.resolve_type_expr(element_type, expr.span)?;
-                if packed_type_size(&elem_ty, &self.packed_classes).is_none() {
-                    return Err(CompileError::new(
-                        expr.span,
-                        "buffer_new<T>() requires a POD scalar, pointer, or packed class element type",
-                    ));
-                }
-                Ok(PhpType::Buffer(Box::new(elem_ty)))
-            }
-            ExprKind::BitNot(inner) => {
-                let ty = self.infer_type(inner, env)?;
-                if !matches!(ty, PhpType::Int | PhpType::Bool | PhpType::False | PhpType::Void) {
-                    return Err(CompileError::new(
-                        expr.span,
-                        "Bitwise NOT requires integer operand",
-                    ));
-                }
-                Ok(PhpType::Int)
-            }
-            ExprKind::NullCoalesce { value, default } => {
-                let vt = self.infer_type(value, env)?;
-                let dt = self.infer_type(default, env)?;
-                let non_null_value = if Self::union_contains_void(&vt) {
-                    self.strip_void_from_union(&vt)
-                } else {
-                    vt
-                };
-                Ok(merge_null_coalesce_result_type(non_null_value, dt))
-            }
-            ExprKind::Pipe { value, callable } => {
-                self.infer_pipe_type(value, callable, expr, env)
-            }
-            ExprKind::Assignment {
-                target,
-                value,
-                result_target,
-                prelude,
-                ..
-            } => {
-                let mut scoped_env = env.clone();
-                self.check_assignment_expression(
-                    target,
-                    value,
-                    result_target.as_deref(),
-                    prelude,
-                    expr.span,
-                    &mut scoped_env,
-                )
-            }
-            ExprKind::ConstRef(name) => {
-                self.constants
-                    .get(name.as_str())
-                    .cloned()
-                    .or_else(|| self.eval_barrier_active.then_some(PhpType::Mixed))
-                    .ok_or_else(|| {
-                        CompileError::new(expr.span, &format!("Undefined constant: {}", name))
-                    })
-            }
-            ExprKind::FirstClassCallable(target) => {
-                self.infer_first_class_callable_target(target, expr.span, env)?;
-                Ok(PhpType::Callable)
-            }
-            ExprKind::Closure {
-                params,
-                variadic,
-                variadic_by_ref,
-                variadic_type: _,
-                return_type,
-                body,
-                is_arrow: _,
-                is_static,
-                captures,
-                capture_refs,
-                by_ref_return: _,
-            } => {
-                if *is_static {
-                    body_must_not_use_this(body, expr.span)?;
-                }
-                self.infer_closure_type(
-                    params,
-                    variadic,
-                    *variadic_by_ref,
-                    return_type,
-                    body,
-                    captures,
-                    capture_refs,
-                    expr,
-                    env,
-                )
-            }
-            ExprKind::Spread(inner) => {
-                let ty = self.infer_type(inner, env)?;
-                match ty {
-                    PhpType::Array(elem_ty) => Ok(*elem_ty),
-                    PhpType::AssocArray { value, .. } => Ok(*value),
-                    _ => Err(CompileError::new(
-                        expr.span,
-                        "Spread operator requires an array",
-                    )),
-                }
-            }
-            ExprKind::NamedArg { value, .. } => self.infer_type(value, env),
-            ExprKind::ClosureCall { var, args } => {
-                self.infer_closure_call_type(var, args, expr, env)
-            }
-            ExprKind::ExprCall { callee, args } => {
-                self.infer_expr_call_type(callee, args, expr, env)
-            }
-            ExprKind::BinaryOp { left, op, right } => {
-                self.infer_binary_op_type(left, op, right, expr, env)
-            }
-            ExprKind::InstanceOf { value, target } => {
-                self.infer_instanceof_type(value, target, expr, env)
-            }
-            ExprKind::NewObject { class_name, args } => {
-                self.infer_new_object_type(class_name.as_str(), args, expr, env)
-            }
-            ExprKind::Clone(inner) => {
-                let ty = self.infer_type(inner, env)?;
-                match ty {
-                    PhpType::Object(class_name) => {
-                        self.check_clone_visibility(&class_name, expr.span)?;
-                        Ok(PhpType::Object(class_name))
-                    }
-                    _ => Err(CompileError::new(expr.span, "clone requires an object value")),
-                }
-            }
-            ExprKind::NewDynamic { name_expr, args } => {
-                // The class is named at runtime; without a literal class
-                // we can't typecheck constructor args or the resulting
-                // object's type. Infer the name expression for its side
-                // effects + warnings, type-check the args generically, and
-                // return Mixed.
-                self.infer_type(name_expr, env)?;
-                for arg in args {
-                    self.infer_type(arg, env)?;
-                }
-                Ok(PhpType::Mixed)
-            }
-            ExprKind::NewDynamicObject {
-                class_name,
-                fallback_class,
-                args,
-                ..
-            } => {
-                let class_ty = self.infer_type(class_name, env)?;
-                if class_ty != PhpType::Str {
-                    return Err(CompileError::new(
-                        class_name.span,
-                        "Dynamic object factory class must be a string",
-                    ));
-                }
-                self.infer_new_object_type(fallback_class.as_str(), args, expr, env)?;
-                self.require_phar_archive_libraries();
-                Ok(PhpType::Object(fallback_class.as_str().to_string()))
-            }
-            ExprKind::PropertyAccess { object, property } => {
-                self.infer_property_access_type(object, property, expr, env)
-            }
-            ExprKind::DynamicPropertyAccess { object, property } => {
-                self.infer_dynamic_property_access_type(object, property, expr, env, false)
-            }
-            ExprKind::NullsafePropertyAccess { object, property } => {
-                self.infer_nullsafe_property_access_type(object, property, expr, env)
-            }
-            ExprKind::NullsafeDynamicPropertyAccess { object, property } => {
-                self.infer_dynamic_property_access_type(object, property, expr, env, true)
-            }
-            ExprKind::StaticPropertyAccess { receiver, property } => {
-                self.infer_static_property_access_type(receiver, property, expr)
-            }
-            ExprKind::MethodCall {
-                object,
-                method,
-                args,
-            } => self.infer_method_call_type(object, method, args, expr, env),
-            ExprKind::NullsafeMethodCall {
-                object,
-                method,
-                args,
-            } => self.infer_nullsafe_method_call_type(object, method, args, expr, env),
-            ExprKind::NullsafeDynamicMethodCall {
-                object,
-                method,
-                args,
-            } => self.infer_nullsafe_dynamic_method_call_type(object, method, args, expr, env),
-            ExprKind::StaticMethodCall {
-                receiver,
-                method,
-                args,
-            } => self.infer_static_method_call_type(receiver, method, args, expr, env),
-            ExprKind::This => self.infer_this_type(expr),
-            ExprKind::PtrCast {
-                target_type,
-                expr: inner,
-            } => self.infer_ptr_cast_type(target_type, inner, expr, env),
-            ExprKind::ClassConstant { receiver } => {
-                self.validate_class_constant_receiver(receiver, expr.span)?;
-                Ok(PhpType::Str)
-            }
-            ExprKind::ObjectClassName { object } => {
-                let object_type = self.infer_type(object, env)?;
-                let object_only = match &object_type {
-                    PhpType::Object(_) => true,
-                    PhpType::Union(members) => {
-                        !members.is_empty()
-                            && members.iter().all(|member| matches!(member, PhpType::Object(_)))
-                    }
-                    _ => false,
-                };
-                if !object_only {
-                    return Err(CompileError::new(
-                        expr.span,
-                        &format!("Cannot use \"::class\" on {}", object_type),
-                    ));
-                }
-                Ok(PhpType::Str)
-            }
-            ExprKind::ScopedConstantAccess { receiver, name } => {
-                self.infer_scoped_constant_access(receiver, name, expr)
-            }
-            ExprKind::NewScopedObject { receiver, args } => {
-                let class_name = match receiver {
-                    crate::parser::ast::StaticReceiver::Self_ => {
-                        self.current_class.clone().ok_or_else(|| {
-                            CompileError::new(
-                                expr.span,
-                                "Cannot use 'new self()' outside a class context",
-                            )
-                        })?
-                    }
-                    crate::parser::ast::StaticReceiver::Static => {
-                        let class_name = self.current_class.clone().ok_or_else(|| {
-                            CompileError::new(
-                                expr.span,
-                                "Cannot use 'new static()' outside a class context",
-                            )
-                        })?;
-                        self.validate_late_bound_constructor_targets(&class_name, args, expr, env)?;
-                        return Ok(PhpType::Object(class_name));
-                    }
-                    crate::parser::ast::StaticReceiver::Parent => {
-                        let current = self.current_class.as_ref().ok_or_else(|| {
-                            CompileError::new(
-                                expr.span,
-                                "Cannot use 'new parent()' outside a class context",
-                            )
-                        })?;
-                        self.classes
-                            .get(current)
-                            .and_then(|info| info.parent.clone())
-                            .ok_or_else(|| {
-                                CompileError::new(
-                                    expr.span,
-                                    &format!("Class '{}' has no parent class", current),
-                                )
-                            })?
-                    }
-                    crate::parser::ast::StaticReceiver::Named(name) => name.as_canonical(),
-                };
-                self.infer_new_object_type(&class_name, args, expr, env)
-            }
-            ExprKind::Yield { key, value } => {
-                if let Some(k) = key {
-                    self.infer_type(k, env)?;
-                }
-                if let Some(v) = value {
-                    self.infer_type(v, env)?;
-                }
-                Ok(PhpType::Mixed)
-            }
-            ExprKind::YieldFrom(inner) => {
-                let inner_ty = self.infer_type(inner, env)?;
-                let supported = match &inner.kind {
-                    ExprKind::ArrayLiteral(_) => true,
-                    ExprKind::FunctionCall { .. } | ExprKind::Variable(_) => {
-                        self.type_accepts(&PhpType::Object("Generator".to_string()), &inner_ty)
-                    }
-                    _ => false,
-                };
-                if !supported {
-                    return Err(CompileError::new(
-                        inner.span,
-                        &format!(
-                            "yield from expects an array literal or Generator, got {:?}",
-                            inner_ty
-                        ),
-                    ));
-                }
-                Ok(PhpType::Mixed)
-            }
-            ExprKind::MagicConstant(_) => {
-                unreachable!("MagicConstant must be lowered before type inference")
-            }
+            ExprKind::IncludeValue { .. }
+            | ExprKind::BoolLiteral(_)
+            | ExprKind::Null
+            | ExprKind::StringLiteral(_)
+            | ExprKind::IntLiteral(_)
+            | ExprKind::FloatLiteral(_)
+            | ExprKind::Variable(_)
+            | ExprKind::Negate(_)
+            | ExprKind::Not(_)
+            | ExprKind::ErrorSuppress(_)
+            | ExprKind::Print(_)
+            | ExprKind::PreIncrement(_)
+            | ExprKind::PreDecrement(_)
+            | ExprKind::PostIncrement(_)
+            | ExprKind::PostDecrement(_)
+            | ExprKind::ArrayLiteralAssoc(_)
+            | ExprKind::Match { .. }
+            | ExprKind::ArrayLiteral(_)
+            | ExprKind::ArrayAccess { .. }
+            | ExprKind::Ternary { .. }
+            | ExprKind::ShortTernary { .. }
+            | ExprKind::Throw(_)
+            | ExprKind::Cast { .. } => self.infer_basic_expr_type(expr, env),
+            ExprKind::FunctionCall { .. }
+            | ExprKind::BufferNew { .. }
+            | ExprKind::BitNot(_)
+            | ExprKind::NullCoalesce { .. }
+            | ExprKind::Pipe { .. }
+            | ExprKind::Assignment { .. }
+            | ExprKind::ConstRef(_)
+            | ExprKind::FirstClassCallable(_)
+            | ExprKind::Closure { .. }
+            | ExprKind::Spread(_)
+            | ExprKind::NamedArg { .. }
+            | ExprKind::ClosureCall { .. }
+            | ExprKind::ExprCall { .. }
+            | ExprKind::BinaryOp { .. }
+            | ExprKind::InstanceOf { .. }
+            | ExprKind::NewObject { .. }
+            | ExprKind::Clone(_)
+            | ExprKind::NewDynamic { .. }
+            | ExprKind::NewDynamicObject { .. }
+            | ExprKind::PropertyAccess { .. }
+            | ExprKind::DynamicPropertyAccess { .. }
+            | ExprKind::NullsafePropertyAccess { .. }
+            | ExprKind::NullsafeDynamicPropertyAccess { .. }
+            | ExprKind::StaticPropertyAccess { .. }
+            | ExprKind::MethodCall { .. }
+            | ExprKind::NullsafeMethodCall { .. }
+            | ExprKind::NullsafeDynamicMethodCall { .. }
+            | ExprKind::StaticMethodCall { .. }
+            | ExprKind::This
+            | ExprKind::PtrCast { .. }
+            | ExprKind::ClassConstant { .. }
+            | ExprKind::ObjectClassName { .. }
+            | ExprKind::ScopedConstantAccess { .. }
+            | ExprKind::NewScopedObject { .. }
+            | ExprKind::Yield { .. }
+            | ExprKind::YieldFrom(_)
+            | ExprKind::MagicConstant(_) =>
+                self.infer_call_or_object_expr_type(expr, env),
         }
     }
-
     /// Returns a variable type, allowing dynamic eval-created locals after an eval barrier.
     fn variable_type_or_eval_dynamic(
         &self,
